melody-agents/requirements.txt
crewai==0.28.8
crewai-tools==0.2.3
langchain-core==0.1.45
langchain-openai==0.0.5
langchain-groq==0.1.5
python-dotenv==1.0.0
streamlit==1.35.0
